backend/requirements.txt
fastapi
uvicorn
pymupdf
google-generativeai